fix: update no submission user list when they do submit
If users submit standup after all the submissions are published, the
list of users who didn't submit the standup gets updated in the Slack
message when they do after the scheduled time. Without this, the list of
users would be stale if they do submit after the scheduled post of
submissions on the channel.

Fixes #25